criado testes de api e unitário, chekin preparatório para iniciar a criação da imagem docker.

+++ Dominio/TaxaJuros/TaxaJuro.cs
@@ -4,8 +4,8 @@
     {
         public TaxaJuro()
         {
-            Taxa = 0.01;
+            Taxa = 0.01m;
         }
-        public double Taxa { get; set; }
+        public decimal Taxa { get; set; }
     }
 }
